Moved single event page tests in to events tests file
